refactor: extract matcher/condition evaluation into hook-matcher.ts
Move the glob/frontmatter matching helpers (globToRegExp, matchesGlob,
matchesFrontmatterFilter) out of hook-manager.ts into a new
src/services/hook-matcher.ts, leaving HookManager as registry + dispatch.
renderPrompt stays in hook-manager.ts — it's prompt-template substitution
(a separate concern) and is imported from there by hook-runner.

Pure code-motion refactor: no behavior changes, no public-surface renames
of the moved functions (they had no external src importers), and the
matcher tests move into test/services/hook-matcher.test.ts.

Sub-issue of #801 (oversized files).

Fixes #1073
