+++ src/services/hook-manager.ts
@@ -12,6 +12,7 @@ import {
 	resolveFeatureToolPolicy,
 } from './feature-definition';
 import { FailurePauseTracker, MAX_CONSECUTIVE_FAILURES } from './failure-pause-tracker';
+import { matchesFrontmatterFilter, matchesGlob } from './hook-matcher';
 
 // ─── Folder / file layout ─────────────────────────────────────────────────────
 
@@ -203,57 +204,12 @@ function validateSlug(raw: string): string {
 	return slug;
 }
 
-// ─── Pure helpers ────────────────────────────────────────────────────────────
-
-/**
- * Compile a glob pattern (`*`, `**`, literal characters) into a RegExp.
- * Single `*` matches any character except `/`; `**` matches any path
- * including `/`. All other regex metacharacters are escaped so user globs
- * cannot accidentally form regex constructs.
- */
-export function globToRegExp(glob: string): RegExp {
-	// Walk the glob once, copying escaped literal characters and emitting
-	// regex equivalents for `**` (matches any path including separators) and
-	// `*` (matches any character except `/`). A single pass avoids the
-	// sentinel-replace approach that needed an unprintable placeholder
-	// character.
-	let pattern = '';
-	for (let i = 0; i < glob.length; i++) {
-		const ch = glob[i];
-		if (ch === '*') {
-			if (glob[i + 1] === '*') {
-				pattern += '.*';
-				i++;
-			} else {
-				pattern += '[^/]*';
-			}
-		} else if (/[.+^${}()|[\]\\]/.test(ch)) {
-			pattern += '\\' + ch;
-		} else {
-			pattern += ch;
-		}
-	}
-	return new RegExp(`^${pattern}$`);
-}
-
-/** Returns true if the path passes the glob (or no glob is provided). */
-export function matchesGlob(path: string, glob: string | undefined): boolean {
-	if (!glob) return true;
-	return globToRegExp(glob).test(path);
-}
-
-/** Returns true if every key in `filter` equals the corresponding frontmatter value. */
-export function matchesFrontmatterFilter(
-	frontmatter: Record<string, unknown> | undefined,
-	filter: Record<string, unknown> | undefined
-): boolean {
-	if (!filter) return true;
-	if (!frontmatter) return false;
-	for (const [key, expected] of Object.entries(filter)) {
-		if (frontmatter[key] !== expected) return false;
-	}
-	return true;
-}
+// ─── Prompt rendering ──────────────────────────────────────────────────────────
+//
+// Glob / frontmatter matching lives in ./hook-matcher (matchesGlob,
+// matchesFrontmatterFilter, globToRegExp). renderPrompt stays here because it's
+// prompt-template substitution, a separate concern from condition evaluation,
+// and is imported by hook-runner from this module.
 
 /** Substitute {{var}} placeholders in `template` from `vars`. */
 export function renderPrompt(template: string, vars: Record<string, string>): string {
