+++ test/services/hook-manager.test.ts
@@ -1,13 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { TFile as MockTFile } from 'obsidian';
-import {
-	HookManager,
-	globToRegExp,
-	matchesGlob,
-	matchesFrontmatterFilter,
-	renderPrompt,
-	type Hook,
-} from '../../src/services/hook-manager';
+import { HookManager, renderPrompt, type Hook } from '../../src/services/hook-manager';
 import { PolicyPreset } from '../../src/types/tool-policy';
 
 // ─── Module mocks ────────────────────────────────────────────────────────────
@@ -163,56 +156,6 @@ function withSeededHooks(plugin: any, hooks: Hook[]): HookManager {
 	(manager as any).initialized = true;
 	return manager;
 }
-
-// ─── Pure helpers ────────────────────────────────────────────────────────────
-
-describe('globToRegExp', () => {
-	it('matches exact paths literally', () => {
-		expect(globToRegExp('Daily/2026-05-04.md').test('Daily/2026-05-04.md')).toBe(true);
-		expect(globToRegExp('Daily/2026-05-04.md').test('Daily/2026-05-05.md')).toBe(false);
-	});
-
-	it('* matches a single segment', () => {
-		expect(globToRegExp('Daily/*.md').test('Daily/2026-05-04.md')).toBe(true);
-		expect(globToRegExp('Daily/*.md').test('Daily/sub/2026-05-04.md')).toBe(false);
-	});
-
-	it('** matches across path separators', () => {
-		expect(globToRegExp('Daily/**/*.md').test('Daily/2026/05/04.md')).toBe(true);
-		expect(globToRegExp('**/notes.md').test('a/b/c/notes.md')).toBe(true);
-	});
-
-	it('escapes regex metacharacters in literals', () => {
-		expect(globToRegExp('a.b+c.md').test('a.b+c.md')).toBe(true);
-		expect(globToRegExp('a.b+c.md').test('aXbYc.md')).toBe(false);
-	});
-});
-
-describe('matchesGlob', () => {
-	it('returns true when no glob is provided', () => {
-		expect(matchesGlob('any/path.md', undefined)).toBe(true);
-	});
-
-	it('applies the compiled glob', () => {
-		expect(matchesGlob('Daily/2026-05-04.md', 'Daily/*.md')).toBe(true);
-		expect(matchesGlob('Notes/2026-05-04.md', 'Daily/*.md')).toBe(false);
-	});
-});
-
-describe('matchesFrontmatterFilter', () => {
-	it('passes when no filter is provided', () => {
-		expect(matchesFrontmatterFilter({ x: 1 }, undefined)).toBe(true);
-	});
-
-	it('rejects when frontmatter is missing but filter requires keys', () => {
-		expect(matchesFrontmatterFilter(undefined, { x: 1 })).toBe(false);
-	});
-
-	it('matches every key/value', () => {
-		expect(matchesFrontmatterFilter({ a: 1, b: 'x' }, { a: 1 })).toBe(true);
-		expect(matchesFrontmatterFilter({ a: 2 }, { a: 1 })).toBe(false);
-	});
-});
 
 describe('HookManager CRUD', () => {
 	function createPluginWithVaultStore() {
